+++ assets/Assets.php
@@ -1,4 +1,5 @@
 <?php
+
 /**
  * @link https://www.humhub.org/
  * @copyright Copyright (c) 2023 HumHub GmbH & Co. KG
@@ -25,13 +26,13 @@ class Assets extends AssetBundle
      * @inheritdoc
      */
     public $css = [
-        'css/sharebetween.min.css'
+        'css/sharebetween.min.css',
     ];
 
     /**
      * @inheritdoc
      */
     public $publishOptions = [
-        'forceCopy' => false
+        'forceCopy' => false,
     ];
 }