+++ messages/ar/base.php
@@ -1,13 +1,19 @@
 <?php
-return array (
-  '<strong>Share</strong> content' => '',
-  'Content not available' => '',
-  'Content you create is automatically displayed on your profile.' => '',
-  'Select Spaces here on which the content is to be additionally displayed.' => '',
-  'Share' => '',
-  'Share this content on your profile stream' => '',
-  'Shared content' => '',
-  'Spaces' => 'الباحات',
-  'This content has either been deleted or you no longer have permission to access it.' => '',
-  '{spaceName} by {userName}' => '',
-);
+
+return [
+    '<strong>Share</strong> content' => '<strong>مشاركة</strong> المحتوى',
+    'Content not available' => 'المحتوى غير متوفر',
+    'Content you create is automatically displayed on your profile.' => 'يتم عرض المحتوى الذي تقوم بإنشائه تلقائيًا في ملفك الشخصي.',
+    'Select Spaces here on which the content is to be additionally displayed.' => 'حدد الباحات هنا التي سيتم عرض المحتوى عليها بشكل إضافي.',
+    'Share' => 'مشاركة',
+    'Share this content on your profile stream' => 'شارك هذا المحتوى في استريم ملفك الشخصي',
+    'Shared content' => 'المحتوى المشارك',
+    'Spaces' => 'الباحات',
+    'This content has either been deleted or you no longer have permission to access it.' => 'تم حذف هذا المحتوى أو لم يعد لديك إذن للوصول إليه.',
+    '{spaceName} by {userName}' => '{spaceName} بواسطة {userName}',
+    'Share Content' => '',
+    'Whenever content (e.g. a post) is shared by a user.' => '',
+    '{user} shared something interesting from Space {space}.' => '',
+    '{user} shared something interesting from dashboard.' => '',
+    '{user} shared something interesting from user {sourceUser}.' => '',
+];