+++ messages/es/base.php
@@ -1,13 +1,19 @@
 <?php
-return array (
-  '<strong>Share</strong> content' => '<strong>Compartir</strong> contenido',
-  'Content not available' => 'Contenido no disponible',
-  'Content you create is automatically displayed on your profile.' => 'El contenido que usted cree será mostrado automáticamente en su perfil.',
-  'Select Spaces here on which the content is to be additionally displayed.' => 'Seleccione aquí Espacios en los cuales este contenido será también mostrado.',
-  'Share' => 'Compartir',
-  'Share this content on your profile stream' => 'Compartir este contenido en la cadena de publicaciones de su perfil',
-  'Shared content' => 'Compartir contenido',
-  'Spaces' => 'Espacios',
-  'This content has either been deleted or you no longer have permission to access it.' => 'Este contenido o ha sido borrado o ya no tiene permisos para accederlo.',
-  '{spaceName} by {userName}' => '{spaceName} de {userName}',
-);
+
+return [
+    '<strong>Share</strong> content' => '<strong>Compartir</strong> contenido',
+    'Content not available' => 'Contenido no disponible',
+    'Content you create is automatically displayed on your profile.' => 'El contenido que usted cree será mostrado automáticamente en su perfil.',
+    'Select Spaces here on which the content is to be additionally displayed.' => 'Seleccione aquí Espacios en los cuales este contenido será también mostrado.',
+    'Share' => 'Compartir',
+    'Share this content on your profile stream' => 'Compartir este contenido en la cadena de publicaciones de su perfil',
+    'Shared content' => 'Compartir contenido',
+    'Spaces' => 'Espacios',
+    'This content has either been deleted or you no longer have permission to access it.' => 'Este contenido o ha sido borrado o ya no tiene permisos para accederlo.',
+    '{spaceName} by {userName}' => '{spaceName} de {userName}',
+    'Share Content' => '',
+    'Whenever content (e.g. a post) is shared by a user.' => '',
+    '{user} shared something interesting from Space {space}.' => '',
+    '{user} shared something interesting from dashboard.' => '',
+    '{user} shared something interesting from user {sourceUser}.' => '',
+];