+++ messages/fr/base.php
@@ -1,13 +1,18 @@
 <?php
-return array (
+return [
   '<strong>Share</strong> content' => '<strong>Partager</strong> du contenu',
   'Content not available' => 'Contenu non disponible',
   'Content you create is automatically displayed on your profile.' => 'Le contenu que vous créez est automatiquement affiché sur votre profil.',
   'Select Spaces here on which the content is to be additionally displayed.' => 'Sélectionnez ici les espaces sur lesquels le contenu doit être affiché en plus.',
   'Share' => 'Partager',
+  'Share Content' => 'Partager le contenu',
   'Share this content on your profile stream' => 'Partagez ce contenu sur le fil d\'actualité de votre profil',
   'Shared content' => 'Contenu partagé',
   'Spaces' => 'Espaces',
   'This content has either been deleted or you no longer have permission to access it.' => 'Ce contenu a été supprimé ou vous n\'êtes plus autorisé à y accéder.',
+  'Whenever content (e.g. a post) is shared by a user.' => 'Dès qu’un contenu (par exemple une publication) est partagé par quelqu’un.',
   '{spaceName} by {userName}' => '{spaceName} par {userName}',
-);
+  '{user} shared something interesting from Space {space}.' => '{user} a partagé quelque chose d’intéressant depuis l’espace {space}.',
+  '{user} shared something interesting from dashboard.' => '{user} a partagé quelque chose d’intéressant depuis la vue d’ensemble.',
+  '{user} shared something interesting from user {sourceUser}.' => '{user} a partagé quelque chose d’intéressant provenant du compte {sourceUser}.',
+];