+++ messages/ja/base.php
@@ -1,13 +1,18 @@
 <?php
-return array (
+return [
   '<strong>Share</strong> content' => 'コンテンツを <strong>共有</strong> します',
   'Content not available' => 'コンテンツは利用できません',
   'Content you create is automatically displayed on your profile.' => '作成したコンテンツはプロフィールに自動的に表示されます。',
   'Select Spaces here on which the content is to be additionally displayed.' => 'ここでコンテンツを追加表示するスペースを選択します。',
   'Share' => '共有',
+  'Share Content' => 'コンテンツを共有する',
   'Share this content on your profile stream' => 'このコンテンツをプロフィール ストリームで共有する',
   'Shared content' => '共有コンテンツ',
   'Spaces' => 'スペース',
   'This content has either been deleted or you no longer have permission to access it.' => 'このコンテンツは削除されているか、アクセスする権限がありません。',
+  'Whenever content (e.g. a post) is shared by a user.' => 'コンテンツ（投稿など）がユーザーによって共有されるたびに。',
   '{spaceName} by {userName}' => '{spaceName}、作成者：{userName}',
-);
+  '{user} shared something interesting from Space {space}.' => '{user}がスペース{space}から面白いものをシェアした。',
+  '{user} shared something interesting from dashboard.' => '{user}がダッシュボードから面白いものをシェアしてくれた。',
+  '{user} shared something interesting from user {sourceUser}.' => '{user}はユーザー{sourceUser}から面白いものを共有しました。',
+];