+++ messages/nl/base.php
@@ -1,13 +1,18 @@
 <?php
-return array (
+return [
   '<strong>Share</strong> content' => 'Inhoud <strong>Delen</strong>',
   'Content not available' => 'Inhoud niet beschikbaar',
   'Content you create is automatically displayed on your profile.' => 'Inhoud die u maakt, wordt automatisch weergegeven op uw profielpagina.',
   'Select Spaces here on which the content is to be additionally displayed.' => 'Selecteer hier de Ruimtes waarop de inhoud ook moet worden weergegeven.',
   'Share' => 'Delen',
+  'Share Content' => 'Inhoud delen',
   'Share this content on your profile stream' => 'Deel deze inhoud op je profielberichten.',
   'Shared content' => 'Gedeelde inhoud',
   'Spaces' => 'Ruimtes',
   'This content has either been deleted or you no longer have permission to access it.' => 'Deze inhoud is verwijderd of u heeft geen toegangsrechten meer.',
+  'Whenever content (e.g. a post) is shared by a user.' => 'Wanneer inhoud (bijvoorbeeld een bericht) door een gebruiker wordt gedeeld.',
   '{spaceName} by {userName}' => '{spaceName} door {userName}',
-);
+  '{user} shared something interesting from Space {space}.' => '{user} heeft iets interessants gedeeld van ruimte {space}.',
+  '{user} shared something interesting from dashboard.' => '{user} heeft iets interessants gedeeld van het dashboard.',
+  '{user} shared something interesting from user {sourceUser}.' => '{user} deelde iets interessants van gebruiker {sourceUser}.',
+];