+++ messages/pl/base.php
@@ -1,13 +1,18 @@
 <?php
-return array (
+return [
   '<strong>Share</strong> content' => '<strong>Podziel się</strong> treścią',
   'Content not available' => 'Zawartość nie jest dostępna',
   'Content you create is automatically displayed on your profile.' => 'Dodana zawartość z automatu trafia do twojego profilu.',
   'Select Spaces here on which the content is to be additionally displayed.' => 'Wybierz strefy, w których zawartość ma zostać wyświetlona.',
   'Share' => 'Udostępnij',
+  'Share Content' => 'Udostępnij zawartość',
   'Share this content on your profile stream' => 'Udostępnij tą zawartość w profilu',
   'Shared content' => 'Współdzielona zawartość',
   'Spaces' => 'Strefy',
   'This content has either been deleted or you no longer have permission to access it.' => 'Ta zawartość została skasowana lub nie masz wystarczających uprawnień dostępu.',
+  'Whenever content (e.g. a post) is shared by a user.' => 'Gdy zawartość (np. wpis) zostanie udostępniona przez użytkownika.',
   '{spaceName} by {userName}' => '{spaceName} przez {userName}',
-);
+  '{user} shared something interesting from Space {space}.' => '{user} udostępnił coś ciekawego z {space}.',
+  '{user} shared something interesting from dashboard.' => '{user} udostępnił coś ciekawego z tablicy.',
+  '{user} shared something interesting from user {sourceUser}.' => '{user} udostępnił coś ciekawego od użytkownika {sourceUser}.',
+];