+++ messages/vi/base.php
@@ -1,13 +1,19 @@
 <?php
-return array (
-  '<strong>Share</strong> content' => '',
-  'Content not available' => '',
-  'Content you create is automatically displayed on your profile.' => '',
-  'Select Spaces here on which the content is to be additionally displayed.' => '',
-  'Share' => '',
-  'Share this content on your profile stream' => '',
-  'Shared content' => '',
-  'Spaces' => 'Các diễn đàn',
-  'This content has either been deleted or you no longer have permission to access it.' => '',
-  '{spaceName} by {userName}' => '',
-);
+
+return [
+    'Spaces' => 'Các diễn đàn',
+    '<strong>Share</strong> content' => '',
+    'Content not available' => '',
+    'Content you create is automatically displayed on your profile.' => '',
+    'Select Spaces here on which the content is to be additionally displayed.' => '',
+    'Share' => '',
+    'Share Content' => '',
+    'Share this content on your profile stream' => '',
+    'Shared content' => '',
+    'This content has either been deleted or you no longer have permission to access it.' => '',
+    'Whenever content (e.g. a post) is shared by a user.' => '',
+    '{spaceName} by {userName}' => '',
+    '{user} shared something interesting from Space {space}.' => '',
+    '{user} shared something interesting from dashboard.' => '',
+    '{user} shared something interesting from user {sourceUser}.' => '',
+];