+++ tests/codeception/_support/FunctionalTester.php
@@ -1,4 +1,5 @@
 <?php
+
 namespace sharebetween;
 
 /**
@@ -19,8 +20,8 @@ namespace sharebetween;
 class FunctionalTester extends \FunctionalTester
 {
     use _generated\FunctionalTesterActions;
-    
-   /**
-    * Define custom actions here
-    */
+
+    /**
+     * Define custom actions here
+     */
 }